Fixed so that pins can be clicked on, and sizes of widgets won't change even as the frame expands.  See details for more.
Found that subviews don't receive mouse events if they are not within the superview's frame. Fixed by expanding the frame when subviews are added. Modified drawing to not use bounds for this reason otherwise it scales too.
Updated tracking area to use this working rect, too, but encountered an issue with mouseEnter/mouseExit because of overlapping subviews, so moved that logic into mouseMoved and detected whether the current view under the mouse is the current one.  Hacky but it works.