Converted the model part of Novella into a framework (see details).
I've shoved Novella into a framework to better separate the code. This is mainly because swift doesn't have namespaces, so I'm relying on the target level (i.e. framework) to restrict things that should be otherwise private. I've made initial changes but there's probably a lot that should be public that's missing.  I'll clean up that in the future. Everything seems to work!