Added a delegate to the story.  See more for details.
Intending to make a delegate that spits out pretty much everything that ever happens to the story.  Currently it handled ADDING of elements globally to the story (but NOT removal - do that some time later), and adding folders and graphs to the Story instance itself.  I need to start forwarding this class to objects so they can call it (or just pass the story and have it say _story.Delegate.x()).  For example, I probably need to prioritize adding it to graphs, so that their addition of nodes (the main action in the program) is handled.  I also need to handle things like link destinations being set etc.