Added KPCTabsControl to Carthage and updated to my branch of JSONSchemaSwift. Also replaced default tab controls with this new one. See more for details.
Partially implemented. Code isn't too clean but it's acceptable. Seems functional but I've not tested throroughly. There's some bugs in their code relating to selection and a big bug with hanging due to mouse events and selection.  If it troubles me to need features where these bugs occur, I will fix and PR it on github.