Added new story delegate func for adding graphs to graphs. See more for details.
NVGraph now has a delegate (most classes eventually will) that is currently called when a subgraph is added, which MainViewController now responds to. When setting the Story's delegate, all existing elements (currently only graphs) have the delegate set. I also set the delegate of graphs upon creation (and will have to do so for other elements in the future).