Set up a dummy test where double clicking a dialog sets the conten of the popup.  See more for details.
The VC content must be set after showing the popover (which is a little counter intuitive) as until shown, the outlets are not configured etc.  I want to impement the delegate for these so that they can be detached individually.  to do that, though I'd need to create a new popup and VC for each one instead of storing them.  This is fine providing they are released from memory upon completion.  I could cache in an array if necesary to store them.