Added a simple toast system built in to NSView.  See details for bug.
Blending using NSVisualEffectView can't have alpha animated, so I opted for a background filter blurring.  Setting the blur to any reasonably high value does indeed blur, but causes artifacting where transparent black rectangles appear on the screen.  Not sure why, so I made it optional until I figure it out.  Blurring also doesn't animate out and must be done outside of  NSAnimationContext, which is awkward.  I either need to not animate it (which is ugly) or time it just right, which would work but I need to get the blur issue fixed first).  Also, when the animation finishes to lower blur to 0, it flicks back to full for one frame, so a delegate must be used to remove all filters upon completion (found on SO).