Work on the condition popover (see more for details).
Wanted to add NSPredicateEditor but that's basically a nogo, so I'm reverting to a syntax highlighted text field for JS code for now.  Using the Highlightr framework.  Tried to get it working with an NSTextArea from IB, but no matter what, setting the container made it fail.  Instead, I have a custom view which I then make an NSTextArea in code and insert into (constraints forcing it in place).  Nothing finished yet but it syntax highlights typed code.  Will probably add buttons on the room for the toolbar at the bottom.