Committing for notes - see more.
Since the snapshot deep copies, all references are now pointing to a graveyard and need updating.  I can do this by recreating the UI elements/reloading them, but it's messy and would not maintain what the user is doing right now.  Maintaining which graphs are open, what's selected etc. when these things are prone to change is a risky thing.  I am going to leave this branch available in case I come up with a good solution or want to hack away at trying it some time.