Basic removal of tab system.  See more for details.
This is the first wave of removing the tab system. Variable and Entity editors are now popup dialogs, although they still need windows attaching to them to be toplevel.  Tab browser has gone and now only one graph is there, but the code is still built for multiple, hence this being the first pass.