A complete rewrite of Novella from the ground up.  See more for details.
Have been working on a separate repository that I am NOT merging, so much work has taken place.  Unfortunate to not merge but it was not suitable to do so.
Major changes involve the model being built-in to the project now, the UI being completely refreshed, and the branch and switch objects being much better.  There's a number of other improvements too but they are too numerous to list.

Almost all features are replicated in this commit as exists in the current commit, and some have been added.

This is officially the first commit of version 2.