Fixed bug where clicking a node would disable its constraints and therefore the bench following, etc.
The method used for bringing a subview to the front involved removing and adding the view which broke its constraints. Saving constraints doesn't work as they become bogus once the view is removed. Manually adding them would be a pain. Instead, I now just do a direct swap in the array which seems to be enough to tell AppKit to behave.